+++ practica_REFAC/src/Customer.java
@@ -1,0 +1,13 @@
+public class Customer {
+        private final String firstName;
+        private final String lastName;
+
+        public Customer(String firstName, String lastName) {
+            this.firstName = firstName;
+            this.lastName = lastName;
+        }
+
+        public String getFullName() {
+            return firstName + " " + lastName;
+        }
+}
