+++ practica_REFAC/src/Order.java
@@ -1,0 +1,20 @@
+import java.util.List;
+
+public class Order {
+    private final List<OrderLineItem> lineItems;
+    private final double taxRate;
+
+    public Order(List<OrderLineItem> lineItems, double taxRate) {
+        this.lineItems = lineItems;
+        this.taxRate = taxRate;
+    }
+
+    public double calculateTotalPrice() {
+        double subtotal = 0.0;
+        for (OrderLineItem item : lineItems) {
+            subtotal += item.getPrice();
+        }
+        double tax = subtotal * taxRate;
+        return subtotal + tax;
+    }
+}
